+++ docs/CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - `Invoke-AmazonTextModel` - minor corrections to debug output and help
     - Cohere
         - `Invoke-CohereCommandRModel` - minor corrections to debug output
+    - Meta
+        - `Invoke-MetaModel` - minor corrections to debug output and help
 - Build changes:
     - Updated bootstrap AWS modules from `4.1.621` to `4.1.690`